+++ protocols/gossipsub/src/rpc.rs
@@ -87,11 +87,15 @@ impl Sender {
             RpcOut::Publish { .. }
             | RpcOut::Graft(_)
             | RpcOut::Prune(_)
+            | RpcOut::Extensions(_)
             | RpcOut::Subscribe(_)
             | RpcOut::Unsubscribe(_) => &self.priority_sender,
-            RpcOut::Forward { .. } | RpcOut::IHave(_) | RpcOut::IWant(_) | RpcOut::IDontWant(_) => {
-                &self.non_priority_sender
-            }
+            RpcOut::Forward { .. }
+            | RpcOut::IHave(_)
+            | RpcOut::IWant(_)
+            | RpcOut::IDontWant(_)
+            | RpcOut::TestExtension
+            | RpcOut::PartialMessage { .. } => &self.non_priority_sender,
         };
         sender.try_send(rpc).map_err(|err| err.into_inner())
     }